Adding Energy_Monitoring_lib
