+++ Code/ESP/WT_Energy_Monitoring_v2/.vscode/c_cpp_properties.json
@@ -10,7 +10,9 @@
             "includePath": [
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/include",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/src",
-                "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib",
+                "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/pubsubclient-2.3/src",
+                "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/Preferences/src",
+                "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/WiFiManager",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/WebServer/src",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/FS/src",
@@ -18,8 +20,6 @@
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/ESPmDNS/src",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/DNSServer/src",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/WiFi/src",
-                "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/pubsubclient-2.3/src",
-                "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/Preferences/src",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/Filters",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/ACS712",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/LoRa/src",
@@ -201,7 +201,6 @@
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/EEPROM",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/RFM69",
                 "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/SPIFlash",
-                "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/ArduinoOTA/src",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/AsyncUDP/src",
                 "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/BLE/src",
@@ -232,7 +231,9 @@
                 "path": [
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/include",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/src",
-                    "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib",
+                    "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/pubsubclient-2.3/src",
+                    "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/Preferences/src",
+                    "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/WiFiManager",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/WebServer/src",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/FS/src",
@@ -240,8 +241,6 @@
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/ESPmDNS/src",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/DNSServer/src",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/WiFi/src",
-                    "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/pubsubclient-2.3/src",
-                    "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/Preferences/src",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/Filters",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/ACS712",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/LoRa/src",
@@ -423,7 +422,6 @@
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/EEPROM",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/RFM69",
                     "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/SPIFlash",
-                    "/home/sarbjeet/gitHub/IoT/Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/ArduinoOTA/src",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/AsyncUDP/src",
                     "/home/sarbjeet/.platformio/packages/framework-arduinoespressif32/libraries/BLE/src",
