+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/myCommon.h
@@ -1,8 +1,10 @@
 #ifndef __MYCOMMON_H__
 #define __MYCOMMON_H__
 
-#include "Communication.h"
+
 #include "receiverBoard.h"
+#include "Communication.h"
+#include "Energy_Monitoring.h"
 
   void getMyMsg();
